feat(legal_compliance): add comprehensive legal compliance management system
