+++ README.md
@@ -1,0 +1,33 @@
+# Legal Compliance Management Module
+
+## Overview
+This Odoo module helps businesses track legal compliance requirements across various departments (HR, Accounting, Manufacturing, etc.). It provides tools to:
+- Manage compliance rules and categories
+- Schedule recurring compliance checks
+- Document compliance evidence
+- Alert responsible persons about upcoming checks
+- Track compliance history
+
+## Features
+- **Rule Management**: Create and categorize compliance rules with different priorities
+- **Recurring Checks**: Set up automatic recurring compliance checks (daily, weekly, monthly, etc.)
+- **Documentation**: Attach supporting documents to rules and checks
+- **Dashboard**: Visual overview of compliance status
+- **Alerts**: Automatic email notifications for upcoming checks
+- **Reporting**: Track compliance history and results
+
+## Installation
+1. Install the module through Odoo Apps
+2. Configure default settings in Settings > Compliance
+3. Create compliance categories
+4. Add your compliance rules
+5. Schedule checks and upload documents
+
+## Dependencies
+- base
+- mail
+- calendar
+- attachment_index
+
+## Support
+For support or feature requests, please contact [SIC Rwanda](https://www.sicrwanda.com/).
