+++ controllers/controllers.py
@@ -1,4 +1,4 @@
-from odoo import http
+from odoo import http, fields
 from odoo.http import request
 
 class ComplianceController(http.Controller):
