+++ models/__init__.py
@@ -1,3 +1,5 @@
-# -*- coding: utf-8 -*-
-
-from . import models
+from . import compliance_category
+from . import compliance_rule
+from . import compliance_check
+from . import compliance_document
+from . import res_config_settings
