+++ models/compliance_rule.py
@@ -5,22 +5,6 @@ import logging
 
 _logger = logging.getLogger(__name__)
 
-class ComplianceCategory(models.Model):
-    _name = 'compliance.category'
-    _description = 'Compliance Category'
-    _order = 'sequence, name'
-
-    name = fields.Char(string='Category Name', required=True, translate=True)
-    sequence = fields.Integer(string='Sequence', default=10)
-    description = fields.Text(string='Description')
-    active = fields.Boolean(string='Active', default=True)
-    rule_ids = fields.One2many('compliance.rule', 'category_id', string='Rules')
-    color = fields.Integer(string='Color Index')
-
-    _sql_constraints = [
-        ('name_uniq', 'unique (name)', 'Category name must be unique!'),
-    ]
-
 class ComplianceRule(models.Model):
     _name = 'compliance.rule'
     _description = 'Compliance Rule'
@@ -179,104 +163,3 @@ class ComplianceRule(models.Model):
             note=f'The compliance check for {self.name} is due on {self.next_check_date}.',
             user_id=self.responsible_id.id
         )
-
-class ComplianceCheck(models.Model):
-    _name = 'compliance.check'
-    _description = 'Compliance Check'
-    _inherit = ['mail.thread', 'mail.activity.mixin']
-    _order = 'check_date desc'
-
-    name = fields.Char(string='Reference', default=lambda self: _('New'), readonly=True)
-    rule_id = fields.Many2one('compliance.rule', string='Rule', required=True, ondelete='cascade')
-    category_id = fields.Many2one(related='rule_id.category_id', string='Category', store=True)
-    department = fields.Selection(related='rule_id.department', string='Department', store=True)
-    check_date = fields.Date(string='Check Date', default=fields.Date.today, required=True)
-    responsible_id = fields.Many2one('res.users', string='Responsible', required=True, default=lambda self: self.env.user)
-    result = fields.Selection([
-        ('compliant', 'Compliant'),
-        ('non_compliant', 'Non-Compliant'),
-        ('partial', 'Partially Compliant'),
-    ], string='Result', required=True)
-    notes = fields.Text(string='Notes')
-    document_ids = fields.One2many('compliance.document', 'check_id', string='Documents')
-    company_id = fields.Many2one('res.company', string='Company', default=lambda self: self.env.company)
-
-    @api.model
-    def create(self, vals):
-        if vals.get('name', _('New')) == _('New'):
-            vals['name'] = self.env['ir.sequence'].next_by_code('compliance.check') or _('New')
-        
-        check = super(ComplianceCheck, self).create(vals)
-        
-        # Update last check date on the rule
-        if check.rule_id:
-            check.rule_id.write({
-                'last_check_date': check.check_date,
-                'alert_sent': False,
-            })
-        
-        return check
-
-class ComplianceDocument(models.Model):
-    _name = 'compliance.document'
-    _description = 'Compliance Document'
-    _inherit = ['mail.thread', 'mail.activity.mixin']
-
-    name = fields.Char(string='Document Name', required=True)
-    rule_id = fields.Many2one('compliance.rule', string='Related Rule', ondelete='cascade')
-    check_id = fields.Many2one('compliance.check', string='Related Check', ondelete='cascade')
-    description = fields.Text(string='Description')
-    file = fields.Binary(string='File', attachment=True)
-    file_name = fields.Char(string='File Name')
-    upload_date = fields.Date(string='Upload Date', default=fields.Date.today)
-    upload_user_id = fields.Many2one('res.users', string='Uploaded By', default=lambda self: self.env.user)
-    expiry_date = fields.Date(string='Expiry Date')
-    is_valid = fields.Boolean(string='Is Valid', compute='_compute_is_valid', store=True)
-
-    @api.depends('expiry_date')
-    def _compute_is_valid(self):
-        today = fields.Date.today()
-        for doc in self:
-            doc.is_valid = not doc.expiry_date or doc.expiry_date >= today
-
-    def action_open_document(self):
-        self.ensure_one()
-        return {
-            'name': self.name,
-            'type': 'ir.actions.act_url',
-            'url': f'/web/content/compliance.document/{self.id}/file/{self.file_name}',
-            'target': 'new',
-        }
-
-class ResConfigSettings(models.TransientModel):
-    _inherit = 'res.config.settings'
-
-    compliance_alert_days = fields.Integer(
-        string='Default Alert Days',
-        config_parameter='legal_compliance.default_alert_days',
-        default=7,
-        help="Default number of days before compliance due date to send alerts"
-    )
-    
-    compliance_manager_id = fields.Many2one(
-        'res.users',
-        string='Compliance Manager',
-        config_parameter='legal_compliance.manager_id',
-        help="Default user responsible for compliance oversight"
-    )
-
-    def set_values(self):
-        super(ResConfigSettings, self).set_values()
-        self.env['ir.config_parameter'].set_param('legal_compliance.default_alert_days', str(self.compliance_alert_days))
-        self.env['ir.config_parameter'].set_param('legal_compliance.manager_id', str(self.compliance_manager_id.id or ''))
-
-    @api.model
-    def get_values(self):
-        res = super(ResConfigSettings, self).get_values()
-        params = self.env['ir.config_parameter'].sudo()
-        
-        res.update(
-            compliance_alert_days=int(params.get_param('legal_compliance.default_alert_days', default=7)),
-            compliance_manager_id=int(params.get_param('legal_compliance.manager_id', default=False)),
-        )
-        return res
